Update e2e 'updatedAt' helper to be more generic ('dateString') helper
